requirements.txt
robotframework>=6.0.0
playwright>=1.35.0
robotframework-browser
robotframework-allure
allure-python-commons
